Move JsNode stuff to separate file in pxtlib
