+++ cli/cli.ts
@@ -2,7 +2,6 @@
 /// <reference path="../built/pxtlib.d.ts"/>
 /// <reference path="../built/pxtcompiler.d.ts"/>
 /// <reference path="../built/pxtsim.d.ts"/>
-/// <reference path="../built/pxtblocks.d.ts" />
 
 (global as any).pxt = pxt;
